Cast for addresses

+++ phase1/pcb.c
@@ -271,7 +271,8 @@ int isPcbValid(pcb_t *p)
 	setSTATUS(status & ~(1 << 3));
 	int a = 1;
 	if (p < pcbTable || p > pcbTable + MAXPROC - 1 ||
-	    searchPcb(&pcbFree_h, p) || (p - pcbTable) % sizeof(pcb_t) != 0)
+	    searchPcb(&pcbFree_h, p) ||
+	    ((int)p - (int)pcbTable) % sizeof(pcb_t) != 0)
 		a = 0;
 	setSTATUS(status);
 	return a;
